+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.3] - 2023-01-03
+
+### Added
+
+- Added `CI` tagging script
+
+### Updated
+
+- Updated `CI` releasing script
+- Updated `CircleCI` config
+- Updated package deployment flow
+
 ## [2.0.2] - 2022-12-10
 
 ### Added